Preliminary cache impl

+++ android-rest/src/main/java/ly/apps/android/rest/cache/ContextPersistentCacheManager.java
@@ -2,24 +2,22 @@ package ly.apps.android.rest.cache;
 
 
 import android.content.Context;
+import com.integralblue.httpresponsecache.compat.MD5;
+import ly.apps.android.rest.utils.Base64;
 import ly.apps.android.rest.utils.IOUtils;
+import ly.apps.android.rest.utils.Logger;
 
 import java.io.File;
-import java.util.Arrays;
+import java.util.Date;
 
-public class ContextPersistenCacheManager implements CacheManager {
+public class ContextPersistentCacheManager implements CacheManager {
 
     private static final String CACHE_EXTENSION = ".obj.cache";
 
     private Context context;
 
-    public ContextPersistenCacheManager(Context context) {
+    public ContextPersistentCacheManager(Context context) {
         this.context = context;
-    }
-
-    @Override
-    public String getCacheKey(Object... args) {
-        return String.valueOf(Arrays.deepHashCode(args));
     }
 
     /**
@@ -30,7 +28,8 @@ public class ContextPersistenCacheManager implements CacheManager {
      * @return the cache file
      */
     private static File getCacheFile(Context context, String cacheKey) {
-        return new File(String.format("%s/%s%s", context.getCacheDir(), cacheKey, CACHE_EXTENSION));
+        String md5 = Base64.encode(cacheKey.getBytes());
+        return new File(String.format("%s/%s%s", context.getCacheDir(), md5, CACHE_EXTENSION));
     }
 
     /**
@@ -38,9 +37,11 @@ public class ContextPersistenCacheManager implements CacheManager {
      */
     @Override
     public void invalidateAll() {
-        for (File file : context.getCacheDir().listFiles()) {
-            if (file.getName().endsWith(CACHE_EXTENSION)) {
-                file.delete();
+        if (context.getCacheDir() != null) {
+            for (File file : context.getCacheDir().listFiles()) {
+                if (file.getName().endsWith(CACHE_EXTENSION)) {
+                    file.delete();
+                }
             }
         }
     }
@@ -48,11 +49,11 @@ public class ContextPersistenCacheManager implements CacheManager {
     /**
      * Adds an object to the cache based on a list of params that are used to calculate the cache key
      *
-     * @param key the cache
-     * @param object  the object to be cached
+     * @param key    the cache
+     * @param object the object to be cached
      */
     @Override
-    public <T> void put(String key, T object) {
+    public <T> void put(String key, T object, CacheInfo cacheInfo) {
         File file = getCacheFile(context, key);
         IOUtils.saveSerializableObjectToDisk(object, file);
     }
@@ -63,16 +64,27 @@ public class ContextPersistenCacheManager implements CacheManager {
      * @return the in memory original object
      */
     @Override
-    public <T> T get(String key) {
+    public <T> T get(final String key, CacheInfo cacheInfo) {
         File file = getCacheFile(context, key);
-        return IOUtils.loadSerializableObjectFromDisk(file);
+        T result = null;
+        long expiresOn = file.lastModified() + cacheInfo.getTimeToLive();
+        long now = new Date().getTime();
+        if (expiresOn >= now) {
+            result = IOUtils.loadSerializableObjectFromDisk(file);
+            Logger.d(String.format("Loading from cache because expiresOn : %d > now : %d", expiresOn, now));
+        } else {
+            invalidate(key);
+            Logger.d(String.format("Invalidated key : %s from cache because expiresOn : %d < now : %d", key, expiresOn, now));
+        }
+        return result;
     }
+
 
     /**
      * Invalidates a cache entry
      */
     @Override
-    public void invalidate(String ...keys) {
+    public void invalidate(String... keys) {
         for (String key : keys) {
             File file = getCacheFile(context, key);
             if (file.exists()) {
@@ -80,4 +92,5 @@ public class ContextPersistenCacheManager implements CacheManager {
             }
         }
     }
+
 }
